+++ samples/multistack-samples/simple-stack-example-multistack-nested-fragment/src/main/java/com/zhuinden/simplestackbottomnavfragmentexample/application/MainActivity.kt
@@ -1,0 +1,38 @@
+package com.zhuinden.simplestackbottomnavfragmentexample.application
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.zhuinden.simplestack.History
+import com.zhuinden.simplestack.SimpleStateChanger
+import com.zhuinden.simplestack.StateChange
+import com.zhuinden.simplestack.navigator.Navigator
+import com.zhuinden.simplestackbottomnavfragmentexample.R
+import com.zhuinden.simplestackbottomnavfragmentexample.features.initial.InitialScreen
+import com.zhuinden.simplestackextensions.fragments.DefaultFragmentStateChanger
+import com.zhuinden.simplestackextensions.services.DefaultServiceProvider
+
+class MainActivity : AppCompatActivity(), SimpleStateChanger.NavigationHandler {
+    private lateinit var fragmentStateChanger: DefaultFragmentStateChanger
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        fragmentStateChanger = DefaultFragmentStateChanger(supportFragmentManager, R.id.container)
+
+        Navigator.configure()
+            .setScopedServices(DefaultServiceProvider())
+            .setStateChanger(SimpleStateChanger(this))
+            .install(this, findViewById(R.id.container), History.of(InitialScreen()))
+    }
+
+    override fun onBackPressed() {
+        if (!Navigator.onBackPressed(this)) {
+            super.onBackPressed()
+        }
+    }
+
+    override fun onNavigationEvent(stateChange: StateChange) {
+        fragmentStateChanger.handleStateChange(stateChange)
+    }
+}