+++ samples/multistack-samples/simple-stack-example-multistack-nested-fragment/src/main/java/com/zhuinden/simplestackbottomnavfragmentexample/features/initial/InitialScreen.kt
@@ -1,0 +1,10 @@
+package com.zhuinden.simplestackbottomnavfragmentexample.features.initial
+
+import androidx.fragment.app.Fragment
+import com.zhuinden.simplestackextensions.fragments.DefaultFragmentKey
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+class InitialScreen : DefaultFragmentKey() {
+    override fun instantiateFragment(): Fragment = InitialFragment()
+}