+++ samples/multistack-samples/simple-stack-example-multistack-nested-fragment/src/main/java/com/zhuinden/simplestackbottomnavfragmentexample/features/root/first/First1Fragment.kt
@@ -1,0 +1,24 @@
+package com.zhuinden.simplestackbottomnavfragmentexample.features.root.first
+
+import android.os.Bundle
+import android.view.View
+import androidx.fragment.app.Fragment
+import com.zhuinden.simplestack.Backstack
+import com.zhuinden.simplestackbottomnavfragmentexample.R
+import com.zhuinden.simplestackbottomnavfragmentexample.databinding.FirstFragment1Binding
+import com.zhuinden.simplestackbottomnavfragmentexample.features.root.RootScreen
+import com.zhuinden.simplestackextensions.fragmentsktx.lookup
+
+class First1Fragment : Fragment(R.layout.first_fragment_1) {
+    private val localStack by lazy { lookup<Backstack>(RootScreen.FIRST_STACK) }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val binding = FirstFragment1Binding.bind(view)
+
+        binding.buttonFirstFragment1GoTo2.setOnClickListener {
+            localStack.goTo(First2Screen())
+        }
+    }
+}