+++ samples/multistack-samples/simple-stack-example-multistack-nested-fragment/src/main/java/com/zhuinden/simplestackbottomnavfragmentexample/features/root/first/First1Screen.kt
@@ -1,0 +1,10 @@
+package com.zhuinden.simplestackbottomnavfragmentexample.features.root.first
+
+import androidx.fragment.app.Fragment
+import com.zhuinden.simplestackextensions.fragments.DefaultFragmentKey
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+class First1Screen : DefaultFragmentKey() {
+    override fun instantiateFragment(): Fragment = First1Fragment()
+}