+++ samples/multistack-samples/simple-stack-example-multistack-nested-fragment/src/main/java/com/zhuinden/simplestackbottomnavfragmentexample/features/root/first/First2Fragment.kt
@@ -1,0 +1,24 @@
+package com.zhuinden.simplestackbottomnavfragmentexample.features.root.first
+
+import android.os.Bundle
+import android.view.View
+import androidx.fragment.app.Fragment
+import com.zhuinden.simplestack.Backstack
+import com.zhuinden.simplestackbottomnavfragmentexample.R
+import com.zhuinden.simplestackbottomnavfragmentexample.databinding.FirstFragment2Binding
+import com.zhuinden.simplestackbottomnavfragmentexample.features.root.RootScreen
+import com.zhuinden.simplestackextensions.fragmentsktx.lookup
+
+class First2Fragment : Fragment(R.layout.first_fragment_2) {
+    private val localStack by lazy { lookup<Backstack>(RootScreen.FIRST_STACK) }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val binding = FirstFragment2Binding.bind(view)
+
+        binding.buttonFirstFragment2GoBack.setOnClickListener {
+            localStack.goBack()
+        }
+    }
+}