+++ samples/multistack-samples/simple-stack-example-multistack-nested-fragment/src/main/java/com/zhuinden/simplestackbottomnavfragmentexample/features/root/second/SecondFragment.kt
@@ -1,0 +1,7 @@
+package com.zhuinden.simplestackbottomnavfragmentexample.features.root.second
+
+import androidx.fragment.app.Fragment
+import com.zhuinden.simplestackbottomnavfragmentexample.R
+
+class SecondFragment : Fragment(R.layout.second_fragment) {
+}