+++ samples/multistack-samples/simple-stack-example-multistack-nested-fragment/src/main/java/com/zhuinden/simplestackbottomnavfragmentexample/features/root/third/ThirdFragment.kt
@@ -1,0 +1,7 @@
+package com.zhuinden.simplestackbottomnavfragmentexample.features.root.third
+
+import androidx.fragment.app.Fragment
+import com.zhuinden.simplestackbottomnavfragmentexample.R
+
+class ThirdFragment : Fragment(R.layout.third_fragment) {
+}