+++ samples/multistack-samples/simple-stack-example-multistack-nested-fragment/src/main/java/com/zhuinden/simplestackbottomnavfragmentexample/features/root/third/ThirdScreen.kt
@@ -1,0 +1,10 @@
+package com.zhuinden.simplestackbottomnavfragmentexample.features.root.third
+
+import androidx.fragment.app.Fragment
+import com.zhuinden.simplestackextensions.fragments.DefaultFragmentKey
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+class ThirdScreen : DefaultFragmentKey() {
+    override fun instantiateFragment(): Fragment = ThirdFragment()
+}